Added heading and altitude controls
Added the heading and altitude controls that allow override and handoff to autopilot. The following keyboard shortcuts are avilable:
ALT+D - Enter a heading value such as 260.
ALT+E - Toggle between manual override and autopilot mode for the heading bug.
ALT+A - Enter an altitude in feet such as 12000.
ALT+L - Toggle between manual override and autopilot mode for the altitude bug.