Added a grid layout to avionics tab.
Added a grid layout to the avionics tab to force a consistent design. The avionics tab has the navigation controls such as speed, altitude, heading, etc. vertically aligned on the left side of the window. The coms devices such as transponder, coms, and navs are vertically aligned on the right side of the window.