Resolved issue with NVDA not reading the gages list when collapsed
Resolved an issue where NVDA wouldn't read the gages list when it was collapsed. To force NVDA to read the list when pressing down arrow, we now force the list of gages to expand when it receives keyboard focus. Likewise, we also force the gages list to collapse when it looses keyboard focus.