Fixed #1, #12, and added setting for default avionics tab
Fixes bug #1, #12, and adds a setting that allows the pilot to choose between two different avionics panels for the autopilot instruments. The first is simplified, and the second is properties. The simplified instrument panel contains a simple interface that focuses on the key aspects of autopilot. The properties version uses a properties grid to display all autopilot instruments. The properties grid version also allows the pilot to sort/catogorize the instruments. No matter the default autopilot instrument panel chosen, control+1 will always take focus to the instruments tab.