Added minimize to tray feature
Added the minimize to tray feature. When the minimize to system tray check box in user interface settings is checked, TFM will minimize to the system tray when the main window is minimized. Otherwise, the main window will work as it currently does. When TFM does minimize to the system tray, you will receive an action center notification stating that TFM minimized itself to the system tray. To restore TFM's main window to the ALT-TAB order, locate the TFM icon in the system tray, then press ENTER on, or double click the icon. The icon is a plane flying over a body of water.