Added ExpressionBuilder class
Added an expression builder that takes care of building expressions and lambdas for the filtering features of the airports and other procedure features. The expression builder can also be used to create formulas such as ground altitude, distances not provided by the FSUIPC C# library, and other general calculations and lambda expressions. For those contributers that might use expression builder for formulas, the expression type is mainly used if the formula or calculation is frequently used, or when the source code needs to know how a particular formula is calculated. Example: If all we care about is the ground speed result (120kn), then expression builder might not be appropriate for ground speed. On the other hand, if TFM needs to know how distance is calculated from waypoint to waypoint because the formula might change at some point during the flight, then expression trees are a good solution since we can modify the distance calculation on the fly during the flight.